pyproject.toml
[tool.poetry]
name = "image-namer"
version = "0.1.0"
description = "Show images and give option to name them"
authors = ["ashariyar <ashariyar@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "image_namer"}]

[tool.poetry.dependencies]
python = "^3.10"
exif = "^1.5.0"
PySimpleGUI = "^4.60.4"
pytesseract = "^0.3.10"
rich = "^13.0.1"
piexif = "^1.1.3"
pillow = "^9.4.0"
filedate = "^2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
